+++ src/xrServerEntities/associative_vector_inline.h
@@ -117,7 +117,7 @@ TEMPLATE_SPECIALIZATION
 //IC	typename _associative_vector::size_type _associative_vector::size						() const
 IC	u32 _associative_vector::size															() const
 {
-	return				(inherited::size());
+	return				((u32)inherited::size());
 }
 
 TEMPLATE_SPECIALIZATION
